fixing wrong statistics, moving gisandapi to differnt module
