+++ c19/api.go
@@ -1,6 +1,6 @@
 /*
  * Covidtron-19000 - a bot for monitoring data about COVID-19.
- * Copyright (C) 2020 Michele Dimaggio, Alessandro Ianne.
+ * Copyright (C) 2020 Michele Dimaggio.
  *
  * Covidtron-19000 is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -28,24 +28,6 @@ import (
 
 	"github.com/thedevsaddam/gojsonq/v2"
 )
-
-func getNazione(nazione string) *GisandData {
-	var data GisandData
-
-	fpath := fmt.Sprintf("%s/gisanddata.json", jsonpath)
-	search := gojsonq.New().
-		File(fpath).
-		WhereContains("country_region", nazione).
-		First()
-
-	if search == nil {
-		return nil
-	} 
-
-	bytes, _ := json.Marshal(search)
-	json.Unmarshal(bytes, &data)
-	return &data
-}
 
 func getAndamento() Andamento {
 	var data Andamento
@@ -129,34 +111,6 @@ func formatTimestamp(timestamp string) string {
 	}
 
 	return tp.Format("15:04 del 02/01/2006")
-}
-
-func GetNazioneMsg(nazione string) string {
-	data := getNazione(nazione)
-
-	if data == nil {
-		log.Println(data)
-		return "Nessun dato disponibile."
-	}
-
-	lastDay := len(data.DailyData) - 1 
-
-	msg := fmt.Sprintf(`*Andamento COVID-19 - %s*
-_Dati aggiornati il: %s_`,
-		data.Country_Region,
-		data.DailyData[lastDay].Date)
-	
-	if data.Province_State != "" {
-		msg = fmt.Sprintf("%s\nStato: %s", msg, data.Province_State)
-	}
-
-	msg = fmt.Sprintf("%s\nGuariti: %d\nDeceduti: %d\nTotale positivi: %d",
-		msg,
-		data.DailyData[lastDay].Recovered,
-		data.DailyData[lastDay].Deaths,
-		data.DailyData[lastDay].Confirmed)
-
-	return msg
 }
 
 func formatNote(codici string) string {
